+++ package/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.1-exp] - 2021-03-29
+- Experimental change: ``EditorPatchProvider`` has public fields for ``PatchThreaded`` (to patch on background thread), ``SuppressExceptions`` and ``EnableException`` (contains last Exception during patching)
+- ``EnablePatch`` returns if patching was successful
+- ``PatchManager`` will attempt patching on main thread when it catches a ``UnityException.MainThread`` exception
+
 ## [1.1.0-pre.4] - 2021-03-24
 - Lazy ``AssetDatabase.FindAssets``, only necessary when path is being accessed via API 
 - Unpatch with explicit fast mode for batch unpatching